+++ src/Arius.Core/Shared/HashCache/Models.cs
@@ -1,9 +1,5 @@
 namespace Arius.Core.Shared.HashCache;
 
-/// <summary>Cheap, platform-provided change signals for one file. See <see cref="SignalSets"/>.</summary>
-[SharedWithinAssembly]
-internal readonly record struct FileChangeSignals(long CtimeTicks, string Inode, string Dev, int SignalSet);
-
 /// <summary>One persisted hashcache row: the cheap signals + sparse fingerprint + cached content hash for a path.</summary>
 internal readonly record struct HashCacheEntry(
     RelativePath Path,
@@ -17,12 +13,3 @@ internal readonly record struct HashCacheEntry(
     int FpAlgo,
     ContentHash ContentHash,
     long LastVerifiedTicks);
-
-/// <summary>Provenance tag stored on a hashcache row so signals are only compared within the same source.</summary>
-[SharedWithinAssembly]
-internal static class SignalSets
-{
-    public const int None    = 0;
-    public const int Posix   = 1;
-    public const int Windows = 2;
-}
